requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.9
pillow==10.2.0
python-decouple==3.8
pandas==2.2.2
matplotlib==3.8.4
